feat: move file to data, core-local, core-remote module
