+++ core/remote/src/main/java/com/emotionstorage/remote/dataSourceImpl/ChatRemoteDataSourceImpl.kt
@@ -1,10 +1,9 @@
-package com.emotionstorage.ai_chat.remote.dataSource
+package com.emotionstorage.remote.dataSourceImpl
 
-import com.emotionstorage.ai_chat.data.dataSource.remote.ChatRemoteDataSource
-import com.emotionstorage.ai_chat.remote.api.ChatApiService
-import com.emotionstorage.ai_chat.remote.response.ChatRoomMessagesResponse
+import com.emotionstorage.data.dataSource.remote.ChatRemoteDataSource
 import com.emotionstorage.domain.common.DataState
 import com.emotionstorage.domain.common.ErrorCode
+import com.emotionstorage.remote.api.ChatApiService
 import com.emotionstorage.remote.response.CustomHttpException
 import java.io.IOException
 import javax.inject.Inject
@@ -61,14 +60,4 @@ class ChatRemoteDataSourceImpl
             } catch (e: Exception) {
                 DataState.Error(e)
             }
-
-
-    override suspend fun getChatRoomMessages(cursor: Long?): DataState<ChatRoomMessagesResponse> =
-        try {
-            val res = chatApiService.getChatRoomMessages(cursor)
-            val data = res.data ?: return DataState.Error(Throwable("getChatRoomMessages(): data is null"))
-            DataState.Success(data)
-        } catch (e: Exception) {
-            DataState.Error(e)
-        }
     }
