+++ core/remote/src/main/java/com/emotionstorage/remote/di/RemoteDataSourceModule.kt
@@ -2,6 +2,8 @@ package com.emotionstorage.remote.di
 
 import com.emotionstorage.data.dataSource.remote.AttendanceRemoteDataSource
 import com.emotionstorage.data.dataSource.remote.AuthRemoteDataSource
+import com.emotionstorage.data.dataSource.remote.ChatRemoteDataSource
+import com.emotionstorage.data.dataSource.remote.ChatWSDataSource
 import com.emotionstorage.data.dataSource.remote.DailyReportRemoteDataSource
 import com.emotionstorage.data.dataSource.remote.FcmRemoteDataSource
 import com.emotionstorage.data.dataSource.remote.GoogleRemoteDataSource
@@ -14,6 +16,8 @@ import com.emotionstorage.data.dataSource.remote.TimeCapsuleRemoteDataSource
 import com.emotionstorage.data.dataSource.remote.UserRemoteDataSource
 import com.emotionstorage.remote.dataSourceImpl.AttendanceRemoteDataSourceImpl
 import com.emotionstorage.remote.dataSourceImpl.AuthRemoteDataSourceImpl
+import com.emotionstorage.remote.dataSourceImpl.ChatRemoteDataSourceImpl
+import com.emotionstorage.remote.dataSourceImpl.ChatWSDataSourceImpl
 import com.emotionstorage.remote.dataSourceImpl.DailyReportRemoteDataSourceImpl
 import com.emotionstorage.remote.dataSourceImpl.FcmRemoteDataSourceImpl
 import com.emotionstorage.remote.dataSourceImpl.GoogleRemoteDataSourceImpl
@@ -82,4 +86,12 @@ abstract class RemoteDataSourceModule {
     @Binds
     @Singleton
     abstract fun bindAttendanceRemoteDataSource(impl: AttendanceRemoteDataSourceImpl): AttendanceRemoteDataSource
+
+    @Binds
+    @Singleton
+    abstract fun bindChatRemoteDataSource(impl: ChatRemoteDataSourceImpl): ChatRemoteDataSource
+
+    @Binds
+    @Singleton
+    abstract fun bindChatWSDataSource(impl: ChatWSDataSourceImpl): ChatWSDataSource
 }
